[FIX] Compiler warnings: src/kernel/modules/navSlices/WMNavSlices.cpp:947: warning: comparison between signed and unsigned integer expressions src/kernel/modules/navSlices/WMNavSlices.cpp:1003: warning: comparison between signed and unsigned integer expressions src/kernel/modules/navSlices/WMNavSlices.cpp:1079: warning: comparison between signed and unsigned integer expressions src/modules/lic/fantom/FMatrix.cpp:757: warning: suggest a space before ‘;’ or explicit braces around empty body in ‘while’ statement src/modules/lic/fantom/FMatrix.cpp:770: warning: suggest a space before ‘;’ or explicit braces around empty body in ‘while’ statement src/modules/lic/fantom/FArray.cpp:227: warning: suggest a space before ‘;’ or explicit braces around empty body in ‘while’ statement src/modules/marchingCubes/WMMarchingCubes.cpp:508: warning: comparison between signed and unsigned integer expressions src/modules/marchingCubes/WMMarchingCubes.cpp:578: warning: comparison between signed and unsigned integer expressions src/modules/marchingCubes/WMMarchingCubes.cpp:659: warning: comparison between signed and unsigned integer expressions src/modules/arbitraryPlane/WMArbitraryPlane.cpp:382: warning: comparison between signed and unsigned integer expressions src/modules/arbitraryPlane/WMArbitraryPlane.cpp:437: warning: comparison between signed and unsigned integer expressions src/modules/arbitraryPlane/WMArbitraryPlane.cpp:505: warning: comparison between signed and unsigned integer expressions src/modules/overlayAtlas/WMOverlayAtlas.cpp:391: warning: comparison between signed and unsigned integer expressions src/modules/overlayAtlas/WMOverlayAtlas.cpp:425: warning: comparison between signed and unsigned integer expressions src/modules/overlayAtlas/WMOverlayAtlas.cpp:497: warning: comparison between signed and unsigned integer expressions src/modules/overlayAtlas/WMOverlayAtlas.cpp:565: warning: comparison between signed and unsigned integer expressions src/modules/splineSurface/WMSplineSurface.cpp:316: warning: comparison between signed and unsigned integer expressions src/modules/splineSurface/WMSplineSurface.cpp:464: warning: comparison between signed and unsigned integer expressions src/modules/splineSurface/WMSplineSurface.cpp:514: warning: comparison between signed and unsigned integer expressions
There are at least those still present:
In file included from /usr/local/cuda/include/cuda_runtime_api.h:53,
   from /home/math/repos/OpenWalnut/src/modules/detTractClustering/WCheckCudaError.cpp:26:
/usr/local/cuda/include/host_defines.h:130:27: warning: anonymous variadic macros were introduced in C99
In file included from /usr/local/cuda/include/builtin_types.h:45,
   from /usr/local/cuda/include/cuda_runtime_api.h:54,
   from /home/math/repos/OpenWalnut/src/modules/detTractClustering/WCheckCudaError.cpp:26:
/usr/local/cuda/include/vector_types.h:73:43: warning: anonymous variadic macros were introduced in C99
/home/math/repos/OpenWalnut/src/common/WSharedLib.cpp: In member function ‘void (* WSharedLib::data::findFunction(const std::string&))()’:
/home/math/repos/OpenWalnut/src/common/WSharedLib.cpp:206: warning: ISO C++ forbids casting between pointer-to-function and pointer-to-object
WMWriteMesh_test.cc:(.text._ZN15WMWriteMeshTest12testSaveZeroEv[WMWriteMeshTest::testSaveZero()]+0xea): warning: the use of `tmpnam' is dangerous, better use `mkstemp'

+++ src/kernel/modules/navSlices/WMNavSlices.cpp
@@ -944,13 +944,13 @@ void WMNavSlices::updateTextures()
         if ( tex.size() > 0 )
         {
             // reset all uniforms
-            for ( int i = 0; i < wlimits::MAX_NUMBER_OF_TEXTURES; ++i )
+            for ( size_t i = 0; i < wlimits::MAX_NUMBER_OF_TEXTURES; ++i )
             {
                 m_typeUniforms[i]->set( 0 );
             }
 
             // for each texture -> apply
-            int c = 0;
+            size_t c = 0;
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
             if ( WKernel::getRunningKernel()->getSelectionManager()->getUseTexture() )
@@ -1076,7 +1076,7 @@ void WMNavSlices::initUniforms( osg::StateSet* rootState )
     m_cmapUniforms.push_back( osg::ref_ptr<osg::Uniform>( new osg::Uniform( "useCmap8", 0 ) ) );
     m_cmapUniforms.push_back( osg::ref_ptr<osg::Uniform>( new osg::Uniform( "useCmap9", 0 ) ) );
 
-    for ( int i = 0; i < wlimits::MAX_NUMBER_OF_TEXTURES; ++i )
+    for ( size_t i = 0; i < wlimits::MAX_NUMBER_OF_TEXTURES; ++i )
     {
         rootState->addUniform( m_typeUniforms[i] );
         rootState->addUniform( m_thresholdUniforms[i] );
